+++ tests/ImageSharp.Tests/Formats/Gif/GifEncoderTests.cs
@@ -1,4 +1,4 @@
-﻿// <copyright file="GifEncoderCoreTests.cs" company="James Jackson-South">
+﻿// <copyright file="GifEncoderTests.cs" company="James Jackson-South">
 // Copyright (c) James Jackson-South and contributors.
 // Licensed under the Apache License, Version 2.0.
 // </copyright>
@@ -10,7 +10,7 @@ namespace ImageSharp.Tests
 
     using ImageSharp.Formats;
 
-    public class GifEncoderCoreTests
+    public class GifEncoderTests
     {
         [Fact]
         public void Encode_IgnoreMetadataIsFalse_CommentsAreWritten()
