+++ tests/ImageSharp.Tests/Formats/Jpg/JpegDecoderTests.cs
@@ -121,5 +121,37 @@ namespace ImageSharp.Tests
                 Assert.Equal(72, image.MetaData.VerticalResolution);
             }
         }
+
+        [Fact]
+        public void Decode_IgnoreMetadataIsFalse_ExifProfileIsRead()
+        {
+            var options = new DecoderOptions()
+            {
+                IgnoreMetadata = false
+            };
+
+            TestFile testFile = TestFile.Create(TestImages.Jpeg.Baseline.Floorplan);
+
+            using (Image image = testFile.CreateImage(options))
+            {
+                Assert.NotNull(image.MetaData.ExifProfile);
+            }
+        }
+
+        [Fact]
+        public void Decode_IgnoreMetadataIsTrue_ExifProfileIgnored()
+        {
+            var options = new DecoderOptions()
+            {
+                IgnoreMetadata = true
+            };
+
+            TestFile testFile = TestFile.Create(TestImages.Jpeg.Baseline.Floorplan);
+
+            using (Image image = testFile.CreateImage(options))
+            {
+                Assert.Null(image.MetaData.ExifProfile);
+            }
+        }
     }
 }
